+++ CHANGELOG.md
@@ -10,11 +10,10 @@
     * Omikron\FactFinder\Shopware6\Export\ExportInterface
         -  added method `getCoveredEntityType` which expect to return a class extending Shopware\Core\Framework\DataAbstractionLayer\Entity which given Exported will be exporting
         -  added method `getProducedExportEntityType` which expect to return a class that implements Omikron\FactFinder\Shopware6\Export\Data\ExportEntityInterface - a exportable version of Entity 
-       
+ - Remove commands for feed exporting (`factfinder:export:products`, `factfinder:export:cms`, `fiactfinder:export:brands`) and replace them with single command `factfinder:data:export`
+   
 ### Add
 - Add Omikron\FactFinder\Shopware6\Export\Data\Factory\FactoryInterface which should be implemented by class which will instantiate object implementing
-- Introduce Category Export
-- Introduce Brand Export
 - Add callback and dom-updated fields to ASN and RecordList CMS Elements which allows to pass inject javascript interacting with Web Components from the PageBuilder
 
 ### Fix